изменил тип переменной Y на double добавил в ресурсы строку с сообщением об ошибке вынес в константы формат даты и времени и сделал нормальное название переменной времени убрал двоение строки вывода за счет оператора объединения ??

+++ My Calc/My Calc/Models/CalcModel.cs
@@ -35,7 +35,7 @@ namespace My_Calc.Models
         /// <summary>
         /// Gets or sets Y
         /// </summary>
-        public int Y { get; set; }
+        public double Y { get; set; }
 
         [Display(Name = "Result", ResourceType = typeof(CalcResources))]
